+++ src/constants/BarrelFiles.ts
@@ -1,0 +1,1 @@
+export const BarrelFiles = ["index.ts", "index.tsx"];