+++ src/models/FileOrFolderToExport.ts
@@ -1,0 +1,4 @@
+export interface FileOrFolderToExport {
+  name: string;
+  type: "file" | "folder";
+}