+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [2.8.0] 03-07-2024
 
+- Updates to use the `vscode.workspace.fs` API
 - [#13](https://github.com/estruyf/vscode-typescript-exportallmodules/issues/13) [#14](https://github.com/estruyf/vscode-typescript-exportallmodules/issues/14): Added support for named exports (thanks to [Mike Rheault](https://github.com/mrheault))
 
 ## [2.7.0] 11-09-2023